% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{acyclic_network}
\alias{acyclic_network}
\title{Create an acyclic network}
\usage{
acyclic_network(size, avg_degree, reduced_root_degree)
}
\arguments{
\item{size}{number of nodes}

\item{avg_degree}{average number of neighbour each node has}

\item{reduced_root_degree}{if true, the degree of the root node is reduced by one}
}
\value{
a network object
}
\description{
Creates an acyclic network with the given size and average degree. Node
degrees follow a Possonian distribution.
}
\seealso{
\code{\link{network_properties}}, \code{\link{network_types}}
}
