% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{barabasialbert_network}
\alias{barabasialbert_network}
\title{Create an Barabasi-Albert network}
\usage{
barabasialbert_network(size, m)
}
\arguments{
\item{size}{number of nodes in the network}

\item{m}{number of nodes each new node attaches to}
}
\value{
a network object
}
\description{
Create an Barabasi-Albert prefential attachment network with the given size and parameter \eqn{m}
}
\seealso{
\code{\link{network_properties}}
}
