% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{brownian_proximity_temporalnetwork}
\alias{brownian_proximity_temporalnetwork}
\title{Create a Brownian proximity network}
\usage{
brownian_proximity_temporalnetwork(
  size,
  avg_degree,
  radius,
  D0,
  D1 = NULL,
  gamma = 0,
  dt = NULL
)
}
\arguments{
\item{size}{the number of nodes in the network}

\item{avg_degree}{the average number of neighbours of each node}

\item{radius}{the contact radius, i.e. the distance below which two nodes are connected by links}

\item{D0}{the diffusivity of non-infected nodes}

\item{D1}{the diffusivity of infected nodes}

\item{gamma}{the exponenent which with diffusivities are scaled as the epidemic grows}

\item{dt}{the time step used when evolving the network}
}
\value{
a network object
}
\description{
Create a Brownian proximity network. \code{size} nodes are places randomly in two
dimensions, and connected by links if their distance does not exceed \code{radius}.
The size of the playing field is chosen such that the expected number of neighbours of
each node is \code{avg_degree}.

The resulting network is \emph{temporal}, i.e. evolves over time. Non-infected nodes diffuse
with diffusivity \eqn{D_0}, infected nodes with diffusivity \eqn{D_1}. As the number of
infected nodes grows, diffusivities are additionally scaled by the factor \eqn{(1 - N_{inf} / N)^{\gamma}}.

The network evolves in descreet time steps of length \code{dt}. If left unspecified, a suitable
\code{dt} is chosen automatically based on the diffusivities and contact radius.
}
\seealso{
\code{\link{network_coordinates}}

\code{\link{network_bounds}}
}
