% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{configmodel_clustered_network}
\alias{configmodel_clustered_network}
\title{Create an network with the specified node degrees and clustering}
\usage{
configmodel_clustered_network(degrees, alpha_or_ck_or_triangles, beta)
}
\arguments{
\item{degrees}{a vector of length \eqn{N} listing the degrees of all nodes}

\item{alpha_or_ck_or_triangles}{TODO}

\item{beta}{TODO}
}
\value{
a network object
}
\description{
Creates a network in nodes have the degrees specified and numbers of trianges specified

TODO: citation
}
\seealso{
\code{\link{network_properties}}, \code{\link{network_types}}
}
