% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{configmodel_network}
\alias{configmodel_network}
\title{Create an network with the specified node degrees}
\usage{
configmodel_network(degrees)
}
\arguments{
\item{degrees}{a vector of length \eqn{N} listing the degrees of all nodes}
}
\value{
a network object
}
\description{
Creates a network in nodes have the degrees specified
}
\seealso{
\code{\link{network_properties}}
}
