% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{cubiclattice_network}
\alias{cubiclattice_network}
\alias{cubiclattice2d_network}
\alias{cubiclattice3d_network}
\alias{cubiclattice4d_network}
\alias{cubiclattice5d_network}
\alias{cubiclattice6d_network}
\alias{cubiclattice7d_network}
\alias{cubiclattice8d_network}
\title{Create an cubic lattice}
\usage{
cubiclattice2d_network(length)

cubiclattice3d_network(length)

cubiclattice4d_network(length)

cubiclattice5d_network(length)

cubiclattice6d_network(length)

cubiclattice7d_network(length)

cubiclattice8d_network(length)
}
\arguments{
\item{length}{the number of nodes on each side of the \eqn{d}-dimensional cube}
}
\value{
a network object
}
\description{
Creates a cubic lattice in the specified number of dimensions
}
\seealso{
\code{\link{network_properties}}, \code{\link{network_types}}
}
