% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{empirical_contact_temporalnetwork}
\alias{empirical_contact_temporalnetwork}
\title{Empirical temporal network comprising short contacts between nodes}
\usage{
empirical_contact_temporalnetwork(
  path,
  dt,
  weight,
  finite_duration = FALSE,
  name = NULL,
  group = "contact",
  gzip = grepl("\\\\.gz$", path),
  download.timeout = 300
)
}
\arguments{
\item{path}{name of the file}

\item{dt}{duration of contacts}

\item{weight}{weight of contacts, instantaneous contacts have effective weight \code{weight} * \code{dt}.}

\item{finite_duration}{whether to treat contacts as having finite duration \code{dt}
or to be instantaneous with effective weight \code{weight} * \code{dt}}

\item{name}{name of a packaged empirical network, see \code{\link{packaged_empirical_network}}}

\item{group}{packaged empirical network group containing the network}

\item{gzip}{whether the file is compressed}

\item{download.timeout}{for packaged networks the download timeout}
}
\description{
Can read networks from a file, or download networks packages for
NEXT-Net directly from the NEXTNet-EmpiricalNetworks repository.

Each line of the file must list a single contact in the form
"\if{html}{\out{<src>}} \if{html}{\out{<dst>}} \if{html}{\out{<time>}}" where \if{html}{\out{<src>}} and \if{html}{\out{<dst>}} are integers specifying the nodes
and \if{html}{\out{<time>}} is a floating-point value. The highest node index appearing within
\if{html}{\out{<src>}} and \if{html}{\out{<dst>}} defines the size of the network. Finite-duration contacts
extend by dt/2 around the time specified in the file.
}
\seealso{
\code{\link{network_properties}}, \code{\link{network_types}}
}
