% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{empirical_weightednetwork}
\alias{empirical_weightednetwork}
\title{Creates a weighted network for an adjacency list stored in a file}
\usage{
empirical_weightednetwork(
  path,
  undirected = TRUE,
  simplify = FALSE,
  idxbase = 1,
  csep = " ",
  wsep = ":",
  gzip = grepl("\\\\.gz$", path)
)
}
\arguments{
\item{path}{name of the file}

\item{undirected}{if \code{TRUE} the network is assumed to be undirected}

\item{simplify}{whether to remove self-edges and multi-edges}

\item{idxbase}{index of the first node (typically 1 or 0, default 1)}

\item{csep}{separator between neighbours/weight pairs, by default whitespace}

\item{wsep}{separator between neighbour and its weight, by default ':'}

\item{gzip}{whether the file is compressed}
}
\value{
a network object
}
\description{
The file must contain one line per node listing first a node and then
pairs of neighbours and their weights. Pairs are separated by whitespace,
neighbours and weights by a colon (':'); other separators can be specifed.
Nodes are identified by numbers starting with \code{idxbase} (by default 1);
the maximal node index that appears in the file defines the size of the
network. For undirected networks (i.e. if \code{undirected=TRUE}), for every
link (u,v) listed in the file the reverse link (v,u) is added as well.

Lines starting with the comment chracter '#' and skipped. If the first
non-comment line does not start with a numerical node index, it is
assumed to be a header line and skipped as well.
}
