% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{erdos_renyi_network}
\alias{erdos_renyi_network}
\alias{erdos_reyni_network}
\title{Create a Erdös-Rényi network}
\usage{
erdos_renyi_network(size, avg_degree)

erdos_reyni_network(size, avg_degree)
}
\arguments{
\item{size}{number of nodes}

\item{avg_degree}{average number of neighbour each node has}
}
\value{
a network object
}
\description{
Creates a Erdös-Rényi network with the given size and average degree
}
\details{
The name of this function was previously miss-spelled \code{erdos_reyni_network},
and it is still available also under its old name.
}
\seealso{
network_properties, network_types
}
