% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{erdos_renyi_temporalnetwork}
\alias{erdos_renyi_temporalnetwork}
\title{Create a temporal Erdös-Rényi network}
\usage{
erdos_renyi_temporalnetwork(size, avg_degree, timescale)
}
\arguments{
\item{size}{number of nodes (\eqn{n})}

\item{avg_degree}{average number of neighbour each node has (\eqn{k})}

\item{timescale}{time scale of network evolution (\eqn{\tau})}
}
\value{
a network object
}
\description{
Creates a temporal Erdös-Rényi network with the given size and average degree.
}
\details{
At any point in time, a temporal Erdös-Rényi network resembles a static
Erdös-Rényi network in that each possible edge exists in the network
with uniform probability \eqn{p_+ = k / (n - 1)} where \eqn{k} as the average
node degree and \eqn{n} the number of nodes. The state of each edge possible
evolves independently over time, appearing with rate \eqn{\lambda_+ = p_+ / \tau}
disappearing with rate \eqn{\lambda_- = p_- / \tau} where \eqn{\tau} is the
time scale.
}
\seealso{
\code{\link{network_properties}}, \code{\link{network_types}}
}
