% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{erdos_renyi_weightednetwork}
\alias{erdos_renyi_weightednetwork}
\title{Create a weighted Erdös-Rényi network}
\usage{
erdos_renyi_weightednetwork(size, avg_degree, weights, probabilities)
}
\arguments{
\item{size}{number of nodes}

\item{avg_degree}{average number of neighbour each node has}

\item{weights}{edge weights}

\item{probabilities}{edge weight probabilities, must be same length as \code{weights}}
}
\value{
a network object
}
\description{
Creates a weighted Erdös-Rényi network with the given size and average degree
and i.i.d. edge weights. The weight distribution is specified by a vector
of weights and a vector of corresponding probabilities.
}
\seealso{
\code{\link{network_properties}}, \code{\link{network_types}}
}
