% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{erdos_reyni_network}
\alias{erdos_reyni_network}
\title{Create a Erdös-Réyni network}
\usage{
erdos_reyni_network(size, avg_degree)
}
\arguments{
\item{size}{number of nodes}

\item{avg_degree}{average number of neighbour each node has}
}
\value{
a network object
}
\description{
Creates a Erdös-Réyni network with the given size and average degree
}
\seealso{
\code{\link{network_properties}}
}
