% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{fully_connected_network}
\alias{fully_connected_network}
\title{Create a fully-connected network}
\usage{
fully_connected_network(size)
}
\arguments{
\item{size}{number of nodes}
}
\value{
a network object
}
\description{
Creates a fully-connected network with the given size
}
\seealso{
\code{\link{network_properties}}
}
