% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{network_properties}
\alias{network_properties}
\alias{network_is_undirected}
\alias{network_size}
\alias{network_outdegree}
\alias{network_neighbour}
\alias{network_adjacencylist}
\alias{network_bounds}
\alias{network_coordinates}
\title{Querying properties of networks}
\usage{
network_is_undirected(nw)

network_size(nw)

network_outdegree(nw, nodes)

network_neighbour(nw, nodes, indices)

network_adjacencylist(nw)

network_bounds(nw)

network_coordinates(nw, nodes)
}
\arguments{
\item{nw}{a network object}

\item{nodes}{vector of node indices}

\item{indices}{vector of neighbour indices}
}
\value{
\itemize{
\item \code{network_is_undirected(nw)}
returns true if the network is not directed, i.e. if there is a link from node \eqn{i}
to \eqn{j} exactly if there is a link from node \eqn{j} to \eqn{i}.
\item \code{network_size(nw)}
returns the number of nodes in the network
\item \code{network_outdegree(nw, node)}
returns a vector of the same length as \code{nodes} containing  the out-degrees (i.e. number
of outgoing links) of the nodes in \code{nodes}.
\item \code{network_neighbour(nw, nodes, indices)}
returns a vector of the same length as \code{nodes} and \code{indices} containing the neighbours
with the given index of the given nodes.
\item \code{network_adjacencylist(nw)}
returns a named list which contains two entries, \emph{nodes} and \emph{neighbours}.
\emph{nodes} contains the indices of all nodes in the network, i.e. \code{1:network_size(nw)}.
\emph{neighbours} is a list of vectors, where \emph{neighbours[i]} lists the neighbours of
node \eqn{i}.
\item \code{network_bounds(nw)}
for networks embedded into \eqn{d}-dimensional space, this function returns a list containing
two vectors of length \eqn{d}, \eqn{x} and \eqn{y}. \eqn{x} is a lower-bound and \eqn{y} the
upper bound for the coordinates of the nodes in the network.
\item \code{network_coordinates(nw, nodes)}
for networks embedded into \eqn{d}-dimensional space, this function returns a \eqn{n\times d}{n x d}
matrix containing the coordinates of the \eqn{n} nodes listed in \code{nodes}.
}
}
\description{
These functions allow properties and the topology of networks to be inspected
}
