% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{network_reproduction_matrix}
\alias{network_reproduction_matrix}
\title{Computes the reproduction matrix \eqn{M_kk}}
\usage{
network_reproduction_matrix(nw)
}
\description{
Computes the reproduction matrix \eqn{M_kk}
}
