% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{packaged_empirical_network}
\alias{packaged_empirical_network}
\title{Downloads a packages empirical network and returns the file path}
\usage{
packaged_empirical_network(
  name,
  group = "undirected",
  format = "gz",
  timeout = 300
)
}
\arguments{
\item{name}{name of the network}

\item{format}{file format}

\item{timeout}{timeout for downloading}

\item{type}{type of network}
}
\value{
path to the downloaded file

Downloaded files are cached in the directory
`rappdirs::user_cache_dir(appname="EmpiricalNetworks", appauthor="NEXTNetR"'
}
\description{
See \href{https://github.com/oist/NEXTNet-EmpiricalNetworks}{NEXTNetR-EmpiricalNetworks}
for a list of available networks.
}
