% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulation}
\alias{simulation}
\title{Create a simulation}
\usage{
simulation(nw, psi, rho = NULL, options = list(), method = "nextreaction")
}
\arguments{
\item{nw}{network to simulate on}

\item{psi}{infection \link[=time_distributions]{time distribution}}

\item{options}{named list of algorithm options, see below}

\item{method}{the simulation method to use, currently only 'nextreaction'}

\item{rhi}{reset/recovery \link[=time_distributions]{time distribution}}
}
\value{
a simulation object. See \code{\link{simulation_run}} for how to
run the simulation and \link{simulation_functions} for other functions that
operate on simulations.
}
\description{
Simulatiojn objects represent a single simulation on a
specific network and using specific tranmission and reset/recovery times.
The network can be \emph{weighted} and/or \emph{temporal}, see \link{network_types}. To
force a temporal network to be treated as a static network, set option
\code{static_network} to \code{TRUE}.
}
\details{
Possible options are
\itemize{
\item \emph{SIR}: Do not make recovered nodes susceptible again as in the SIR model. Default FALSE.
\item \emph{static_network}: Treat network as static, even if it is a temporal network. Default FALSE.
\item \emph{shuffle_neighbours}: Shuffle the neighbours upon infecting a node. Default TRUE.
\item \emph{edges_concurrent}: Activate all outgoing edges simultaenously or sequentially. If set to true, neighbours are implicitly shuffled and \emph{shuffle_neighbours} thus has no effect. Default FALSE.
}
}
\examples{
# Create contact network network
nw <- erdos_renyi_network(1e5, 5)
# Create transmission and reset time distributions
psi <- lognormal_time(6, 30, 0.1)
rho <- weibull_time(shape=5, scale=50)
# Create simulation and specifiy initial set of infections
sim <- simulation(nw, psi, rho)
simulation_addinfections(sim, nodes=c(1), times=c(0.0))
# Run simulation until time t=100 or 200,000 infections have occured
r <- simulation_run(sim, stop=list(time=300, total_infected=300e3))
# Plot the number of infected nodes against time
plot(r$time, r$infected, type='l')
}
\seealso{
\code{\link{simulation_functions}}
}
