% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulation_run}
\alias{simulation_run}
\title{Runs a simulation until a stopping condition occurs}
\usage{
simulation_run(sim, stop, opts = list())
}
\arguments{
\item{sim}{a simulation object}

\item{stop}{a list of stopping conditions}

\item{opts}{return value options}
}
\value{
a \code{data.frame} containing the columns
\itemize{
\item \emph{time}: time since the start of the simulation
\item \emph{epidemic_step}: total number of epidemic events so far
\item \emph{network_step}: total number of network events so far
\item \emph{kind}: kind of event (\emph{outside_infection}, \emph{infection}, \emph{reset},
\emph{neighbour_added}, \emph{neibhour_removed}, \emph{instantenous_contact})
\item \emph{node}: node affected by the event
\item \emph{neighbour}: neighbour (for \emph{infection}, \emph{neighbour_added}, \emph{neibhour_removed})
\item \emph{total_infected}: total number of infected nodes since simulations start
\item \emph{total_reset}: total number of recovered/reset nodes since simulation start
\item \emph{infected}: current number of infected nodes
}
}
\description{
Runs the specified simulation until a stopping condition occurs
and returns a table listing all events that occurred. \code{simulation_run} can be
called repeatedly and will continue from the point where the last run stopped.
Note that for an epidemic to start, \emph{initially infected} nodes must be defined
with \code{\link{simulation_addinfections}}.
}
\details{
Stopping conditions are specified as a named list of threshold values for
any combination of the following variables:
\itemize{
\item \emph{epidemic_steps}: Number of epidemic events (infection, outside_infection,
reset) since the simulation was started.
\item \emph{network_steps}:  Number of network events (edge added, edge removed,
instantaneous contact) since the simulation was started.
Only relevant for simulations on dynamic networks.
\item \emph{time}:           time since the simulation was started
\item \emph{infected}:       Number of currently infected nodes
\item \emph{total_infected}: Total number of infections since the start of the simulation
\item \emph{total_reset}:    Total number of recoveries/resets since the start of the simulation
}

Possible return value options are:
\itemize{
\item \emph{network_events}: Include network events in the returned \code{data.frame}. Default \emph{false}
\item \emph{epidemic_events}: Include epidemic events in the returned \code{data.frame}. Default \emph{true}
}
}
\examples{
# Create contact network network
g <- erdos_renyi_network(1e5, 5)
# Create transmission and reset time distributions
psi <- lognormal_time(3, 30)
rho <- lognormal_time(50, 20)
# Create simulation and specifiy initial set of infections
sim <- nextreaction_simulation(g, psi, rho)
simulation_addinfections(sim, nodes=c(1), times=c(0.0))
# Run simulation until time t=100 or 200,000 infections have occured
r <- simulation_run(sim, list(time=100, total_infected=2e5))
# Plot the number of infected nodes against time
plot(r$time, r$infected, type='l')
}
\seealso{
\code{\link{simulation_functions}}, \code{\link{nextreaction_simulation}},
\code{\link{nmga_simulation}}
}
