% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulation_step}
\alias{simulation_step}
\title{Outdated version of \code{simulation_run}}
\usage{
simulation_step(sim_, steps, opts = list())
}
\description{
This is a compatibility wrapper for \code{simulation_run}.
}
\seealso{
\code{\link{simulation_run}}, \code{\link{simulation_functions}}
}
