% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulation_types}
\alias{simulation_types}
\title{Creating simulations

TODO: Writeme

\link{nextreaction_simulation}
\link{nextreaction_simulation_meanfield}
\link{nmga_simulation}}
\description{
Creating simulations

TODO: Writeme

\link{nextreaction_simulation}
\link{nextreaction_simulation_meanfield}
\link{nmga_simulation}
}
