% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission_time.R
\name{time_functions}
\alias{time_functions}
\alias{time_sample}
\alias{time_density}
\alias{time_hazardrate}
\alias{time_survivalprobability}
\alias{time_survivalquantile}
\title{Functions that operate on transmission/recovery/reset time distributions}
\usage{
time_sample(n, timedistribution, t = 0, m = 1)

time_density(timedistribution, tau)

time_hazardrate(ttimedistributiontr, tau)

time_survivalprobability(timedistribution, tau, t = 0, m = 1)

time_survivalquantile(timedistribution, p, t = 0, m = 1)
}
\arguments{
\item{n}{number of samples to draw}

\item{timedistribution}{a \link[=time_distributions]{time_distribution}.}

\item{t}{the conditioning time}

\item{m}{multiplicity}

\item{tau}{non-negative time}
}
\value{
\itemize{
\item \code{time_sample(n, timedistribution, t, m)}. Samples \code{n} values from
distribution \eqn{\Psi_{t,m}}.
\item \code{time_density(timedistribution, tau)}. Evaluates the density at points
\code{tau} of distribution \eqn{\Psi_{t,m}}.
\item \code{time_hazardrate(timedistribution, tau)}. Evaluates the hazardrate
\eqn{\lambda(\tau)} of the \emph{base} distribution \eqn{\Psi}. The hazard
rate of \eqn{\Psi_{t,m}} is simply \eqn{m \times \lambda(\tau)}.
\item \code{time_survivalprobability(timedistribution, tau, t, m)}. Evaluates the
survivalfunction \eqn{\Psi_{t,m}} at points \code{tau}.
\item \code{time_survivalquantile(timedistribution, p, t, m)}. Computes the
\eqn{p}-quantiles of distribution \eqn{\Psi_{t,m}}.
}
}
\description{
A time distributions represents a family of distributions parametrized by
a conditioning time \eqn{t \geq 0}, a multiplicity \eqn{m > 0} over
a base distribution with survival function
\eqn{\Psi(\tau) = e^{-\lambda(\tau)}}. The parameters represent (i)
conditioning of the base distribution on values \eqn{\geq t} and (ii)
taking the \emph{minimum} over \eqn{m} i.i.d. samples of \eqn{\Psi}. The survival
function of \eqn{\Psi_{t,m}} is therefore \eqn{(\Psi(\tau)/\Psi(t))^m}.
}
\seealso{
\code{\link{time_distributions}}
}
