% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{wattsstrogatz_network}
\alias{wattsstrogatz_network}
\title{Create an Watts-Strogatz network}
\usage{
wattsstrogatz_network(size, k, p)
}
\arguments{
\item{size}{number of nodes in the network}

\item{k}{parameter k}

\item{p}{parameter p}
}
\value{
a network object
}
\description{
Create an Watts-Strogatz  network with the given size and parameters \eqn{k}, \eqn{p}
}
\seealso{
\code{\link{network_properties}}
}
