% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{activity_driven_temporalnetwork}
\alias{activity_driven_temporalnetwork}
\title{TODO}
\usage{
activity_driven_temporalnetwork(
  activities,
  m,
  eta,
  b,
  eta_inf = eta,
  b_inf = b
)
}
\description{
TODO
}
