% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{adjacencylist_weightednetwork}
\alias{adjacencylist_weightednetwork}
\title{Create a network from an adjacency list}
\usage{
adjacencylist_weightednetwork(
  adjacencylist,
  is_undirected = FALSE,
  above_diagonal = is_undirected
)
}
\arguments{
\item{adjacencylist}{a list of vectors containing the neighbours of each node. Same format as the return value of \code{\link{weighted_network_adjacencylist}}.}

\item{is_undirected}{\code{TRUE} if the network is supposed to be undirected, i.e. contains a link from \eqn{i} to \eqn{j} exactly if it contains a link from \eqn{j} to \eqn{i}.}

\item{above_diagonal}{set to \code{TRUE} if the network is undirected and the adjacencylist only contains edges \eqn{i,j} with \eqn{i \leq j}, i.e. represent the upper triangular submatrix of the adjacency matrix. Defaults to \code{TRUE} for undirected networks.}
}
\value{
a network object
}
\description{
Create an network object from an adjacency list
}
\seealso{
\code{\link{network_properties}}, \code{\link{network_types}}
}
