% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{empirical_network}
\alias{empirical_network}
\title{Creates a network for an adjacency list stored in a file}
\usage{
empirical_network(filename)
}
\arguments{
\item{filename}{name of the file}
}
\value{
a network object
}
\description{
The file must contain one line per node listing that node's neighbours
separated by commas. Nodes are referred to by their line numbers which
start with zero.
}
