% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{network_properties}
\alias{network_properties}
\alias{network_is_undirected}
\alias{network_is_simple}
\alias{network_is_weighted}
\alias{network_is_temporal}
\alias{network_size}
\alias{network_outdegree}
\alias{network_neighbour}
\alias{network_neighbour_weight}
\alias{network_adjacencylist}
\alias{weighted_network_adjacencylist}
\alias{network_bounds}
\alias{network_coordinates}
\title{Querying properties of networks}
\usage{
network_is_undirected(nw)

network_is_simple(nw)

network_is_weighted(nw)

network_is_temporal(nw)

network_size(nw)

network_outdegree(nw, nodes)

network_neighbour(nw, nodes, indices)

network_neighbour_weight(nw, nodes, indices)

network_adjacencylist(nw, above_diagonal = network_is_undirected(nw))

weighted_network_adjacencylist(nw, above_diagonal = network_is_undirected(nw))

network_bounds(nw)

network_coordinates(nw, nodes)
}
\arguments{
\item{nw}{a network object}

\item{nodes}{vector of node indices}

\item{indices}{vector of neighbour indices}
}
\value{
\itemize{
\item \code{network_is_simple(nw)}
returns true if the network is simple, i.e. if it neither contains
self-edges nor multi-edges. Note that \code{false} does not imply that the
network necessarily contains self- or multi-edges, only that is is not
guaranteed to be simple by construction.
\item \code{network_is_undirected(nw)}
returns true if the network is not directed, i.e. if there is a link from node \eqn{i}
to \eqn{j} exactly if there is a link from node \eqn{j} to \eqn{i}. Note that
similar to \code{network_is_simple}, a return value of \code{true} guarantees that the
network is undirected, but \code{false} does not imply the existence of an edge
without a reversed counterpart.
\item \code{network_is_weighted(nw)}
return true if the network is weighted, i.e. if links have associated weights
\item \code{network_is_temporal(nw)}
return true if the network is temporal, i.e. if links can appear and disappear
during epidemic simulations
\item \code{network_size(nw)}
returns the number of nodes in the network
\item \code{network_outdegree(nw, node)}
returns a vector of the same length as \code{nodes} containing  the out-degrees (i.e. number
of outgoing links) of the nodes in \code{nodes}.
\item \code{network_neighbour(nw, nodes, indices)}
returns a vector of the same length as \code{nodes} and \code{indices} containing the neighbours
with the given index of the given nodes.
\item \code{network_neighbour_weight(nw, nodes, indices)}
returns a named list containing vectors "n" and "w". "n" is a vector of the same
length as \code{nodes} and \code{indices} and contains the neighbours with the given index of the
given nodes. "w" is a vector of the same length and contains the corresponding weights.
\item \code{network_adjacencylist(nw, above_diagonal=network_is_undirected(nw))}
returns a list of interger vectors where \eqn{i}-th vector contains the neighbours of
node \eqn{i}. For undirected networks, of the edge pair \eqn{(i,j)} and \eqn{(j, i)}
only the edge \eqn{(i,j)} with \eqn{i \leq j} is output, i.e. only the upper triangular
submatrix of an symmetric adjacency matrix is considered. To include both edges,
set \code{above_diagonal=FALSE}.
\item \code{weighted_network_adjacencylist(nw, above_diagonal=network_is_undirected(nw))}
returns a list of named lists containing vectors "n" and "w". The vector "n"
in the \eqn{i}-th list contains neighbours of node \eqn{i}, and the vector "w" contains the
corresponding weights. For undirected networks, of the edge pair \eqn{(i,j)} and \eqn{(j, i)}
only the edge \eqn{(i,j)} with \eqn{i \leq j} is output, i.e. only the upper triangular
submatrix of an symmetric adjacency matrix is considered. To include both edges,
set \code{above_diagonal=FALSE}.
\item \code{network_bounds(nw)}
for networks embedded into \eqn{d}-dimensional space, this function returns a list containing
two vectors of length \eqn{d}, \eqn{x} and \eqn{y}. \eqn{x} is a lower-bound and \eqn{y} the
upper bound for the coordinates of the nodes in the network.
\item \code{network_coordinates(nw, nodes)}
for networks embedded into \eqn{d}-dimensional space, this function returns a \eqn{n\times d}{n x d}
matrix containing the coordinates of the \eqn{n} nodes listed in \code{nodes}.
}
}
\description{
These functions allow properties and the topology of networks to be inspected
}
\seealso{
\code{\link{network_types}}
}
