% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{network_types}
\alias{network_types}
\title{Creating networks}
\description{
NEXTNetR currently supports the following types of static, unweighted networks
\itemize{
\item \link{empirical_network}
\item \link{erdos_renyi_network}
\item \link{fully_connected_network}
\item \link{acyclic_network}
\item \link{configmodel_network}
\item \link{configmodel_clustered_network}
\item \link{wattsstrogatz_network}
\item \link{barabasialbert_network}
\item \link{cubiclattice_network}
\item \link{adjacencylist_network}
}

the following static weighted networks
\itemize{
\item \link{weighted_adjacencylist_network}
}

and following temporal networks
\itemize{
\item \link{empirical_temporalnetwork}
\item \link{brownian_proximity_temporalnetwork}
\item \link{sirx_temporalnetwork}
}
}
