% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{nextreaction_simulation}
\alias{nextreaction_simulation}
\title{Create a simulator using the NextReaction algorithm}
\usage{
nextreaction_simulation(nw, psi, rho = NULL, options = list())
}
\description{
Possible options are
shuffle_neighbours: Whether to shuffle the neighbours upon infecting a node
}
