% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{nmga_simulation}
\alias{nmga_simulation}
\title{Create a simulator using the non-Markovian Gillespie (nMGA) algorithm}
\usage{
nmga_simulation(nw, psi, rho = NULL, options = list())
}
\description{
Possible options are
approx_threshold: Threshold for infected nodes at which the approximate algorithm is used
max_dt: Maximum timestep allowed for the approximate algorithm
tauprec: Numerical precision used to invert the CDF in the exact algorithm
}
