% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission_time.R
\name{time_distributions}
\alias{time_distributions}
\alias{exponential_time}
\alias{lognormal_time}
\alias{gamma_time}
\alias{weibull_time}
\alias{polynomial_rate_time}
\alias{infectiousness_time}
\alias{deterministic_time}
\title{Creating time distributions}
\usage{
exponential_time(lambda, p_infinity = 0)

lognormal_time(mean, var, p_infinity = 0)

gamma_time(mean, var, p_infinity = 0)

weibull_time(shape, scale, p_infinity = 0)

polynomial_rate_time(coeffs)

infectiousness_time(tau, lambda)

deterministic_time(tau)
}
\value{
\itemize{
\item \code{exponential_time(lambda, p_infinity)}. Returns a time distribution representing
an exponential distribution with rate \code{lambda} which in addition to finite
values produces the value infinity with probability \code{p_infinity}. This represents
the case of constant infectiousness \eqn{\lambda(\tau)=\lambda}.
\item \code{lognormal_time(mean, var, p_infinity)}. Returns a time distribution representing
a Log-normal distribution with the given \code{mean} and \code{variance}, which in addition
to finite  values produces the value infinity with probability \code{p_infinity}.
\item \code{gamma_time(mean, var, p_infinity)}. Returns a time distribution representing
a Gamma distribution with the given \code{mean} and \code{variance}, which in addition
to finite  values produces the value infinity with probability \code{p_infinity}.
\item \code{weibull_time(mean, var, p_infinity)}. Returns a time distribution representing
a Weibull distribution with the given \code{shape} and \code{scale} parameter, which in addition
to finite  values produces the value infinity with probability \code{p_infinity}.
The distribution has mean \eqn{b \Gamma(1 + 1/a)} and variance
\eqn{b^2(\Gamma(1 + 2/a) - \Gamma^2(1 + 1/a))}for shape \eqn{a} and scale \eqn{b}.
\item \code{polynomial_rate_time(coeffs)}. Distribution with survival function
\eqn{\Psi(\tau) = e^{-p(\tau)}} for a polynomial infectiousness (hazard rate)
\eqn{p = c[1] + c[2] x + c[3] x^2 + \ldots} with non-negative coefficients.
\item \code{infectiousness_time(tau, lambda)}. Distribution defined in term of the infectiousness
function \eqn{\lambda(\tau)} specified for discrete points \eqn{\tau_i, \lambda_i=\lambda(\tau_i)}
through vectors \code{tau} and \code{lambda}. The vectors must have the same length and be non-empty. In
between the specified points, \eqn{\lambda(\tau)} is interpolated linearly. After the largest
specified \eqn{\tau_i}, the infectiousness \eqn{\lambda(\tau)} is assumed to be constant.
\item \code{deterministic_time(tau)}. Deterministic time with fixed value \code{tau}
}
}
\description{
NEXTNetR supports a series of pre-defined distributions for transmission and
recover/reset time, and also allows custom distributions to be defined
with \link{userdefined_time}. For \code{exponential_time}, \code{lognormal_time}, \code{gamma_time},
and \code{weibull_time} the names reflect the shape of the transmission time density.
Except for \code{exponential_time}, all these distributions have two parameters, and
are parametrized by their mean and variance. The additional parameter \eqn{p_\infty}
adds infinite times as a possible outcome with the specified probability, i.e
scales the density by \eqn{1-p_\infty} so that it is no longer normalized.
Instead of by their transmission time density, time distributions can alternatively
be thought of as being defined by their infectiousness function \eqn{\lambda(\tau)},
this is the case for \code{polynomial_rate_time} and \code{infectiousness_time}. In this case,
\eqn{p_\infty} is implicity defined by \eqn{\lambda(\tau)}. See the discussion in
\link{time_functions} for how the two represenations are related.

Each \code{time_distribution} object actually represents a two-parameer family
of distributions, see \link{time_functions} for a full discussion and for functions
that operate on time distribution objects.
}
\seealso{
\code{\link{time_functions}}
}
