% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission_time.R
\name{userdefined_time}
\alias{userdefined_time}
\title{User-defined time distributions}
\usage{
userdefined_time(
  survival,
  density,
  sample = NULL,
  quantile = NULL,
  survival_nargs,
  sample_nargs = NULL,
  quantile_nargs = NULL,
  p_infinity
)
}
\description{
TODO
}
\seealso{
\code{\link{time_distributions}}, \code{\link{time_functions}}
}
