% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission_time.R
\name{userdefined_time}
\alias{userdefined_time}
\title{User-defined time distributions}
\usage{
userdefined_time(
  sample,
  sample_is_trinary,
  survival,
  survival_is_trinary,
  density,
  survivalquantile,
  quantile_is_trinary,
  p_infinity
)
}
\description{
TODO
}
\seealso{
\code{\link{time_distributions}}, \code{\link{time_functions}}
}
